+++ index.js
@@ -1,7 +1,7 @@
-/** @format */
+import { AppRegistry } from 'react-native';
+import { name as appName } from './app.json';
+import { HomePage } from "./src/pages/home-page/home-page";
 
-import {AppRegistry} from 'react-native';
-import App from './App';
-import {name as appName} from './app.json';
-
-AppRegistry.registerComponent(appName, () => App);
+AppRegistry.registerComponent(appName, () => {
+	return HomePage;
+});
